requirement.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
decorator==5.1.1
Django==5.1.4
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.1.1
mutagen==1.47.0
numpy==2.2.1
pillow==10.4.0
proglog==0.1.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
tzdata==2024.2
urllib3==2.3.0
